+++ core/node/notifications/service.go
@@ -2,8 +2,6 @@ package notifications
 
 import (
 	"context"
-	"encoding/hex"
-	"sync"
 	"time"
 
 	"connectrpc.com/connect"
@@ -11,6 +9,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 
 	"github.com/river-build/river/core/config"
+	"github.com/river-build/river/core/node/auth"
 	. "github.com/river-build/river/core/node/base"
 	"github.com/river-build/river/core/node/crypto"
 	"github.com/river-build/river/core/node/events"
@@ -26,17 +25,15 @@ import (
 
 type (
 	Service struct {
-		notificationsConfig           config.NotificationsConfig
-		onChainConfig                 crypto.OnChainConfiguration
-		userPreferences               UserPreferencesStore
-		riverRegistry                 *registries.RiverRegistryContract
-		nodes                         []nodes.NodeRegistry
-		listener                      events.StreamEventListener
-		streamsTracker                *notificationssync.StreamsTracker
-		metrics                       infra.MetricsFactory
-		pendingAuthenticationRequests sync.Map
-		sessionTokenSigningKey        any
-		sessionTokenSigningAlgo       string
+		auth.AuthServiceMixin
+		notificationsConfig config.NotificationsConfig
+		onChainConfig       crypto.OnChainConfiguration
+		userPreferences     UserPreferencesStore
+		riverRegistry       *registries.RiverRegistryContract
+		nodes               []nodes.NodeRegistry
+		listener            events.StreamEventListener
+		streamsTracker      *notificationssync.StreamsTracker
+		metrics             infra.MetricsFactory
 	}
 )
 
@@ -63,41 +60,21 @@ func NewService(
 		return nil, err
 	}
 
-	// set defaults
-	if notificationsConfig.Authentication.ChallengeTimeout <= 0 {
-		notificationsConfig.Authentication.ChallengeTimeout = 30 * time.Second
+	service := &Service{
+		notificationsConfig: notificationsConfig,
+		onChainConfig:       onChainConfig,
+		userPreferences:     userPreferences,
+		riverRegistry:       riverRegistry,
+		nodes:               nodes,
+		listener:            listener,
+		streamsTracker:      tracker,
+		metrics:             metrics,
 	}
-	if notificationsConfig.Authentication.SessionToken.Lifetime <= 0 {
-		notificationsConfig.Authentication.SessionToken.Lifetime = 30 * time.Minute
-	}
-
-	if len(notificationsConfig.Authentication.SessionToken.Key.Key) != 64 {
-		return nil, RiverError(Err_BAD_CONFIG, "Invalid session token key length",
-			"len", len(notificationsConfig.Authentication.SessionToken.Key.Key)).
-			Func("NewService")
-	}
-
-	key, err := hex.DecodeString(notificationsConfig.Authentication.SessionToken.Key.Key)
-	if err != nil {
-		return nil, RiverError(Err_BAD_CONFIG, "Invalid session token key (not hex)").Func("NewService")
+	if err := service.AuthServiceMixin.Init(&notificationsConfig.Authentication); err != nil {
+		return nil, err
 	}
 
-	if len(key) != 32 {
-		return nil, RiverError(Err_BAD_CONFIG, "Invalid session token key decoded length").Func("NewService")
-	}
-
-	return &Service{
-		notificationsConfig:     notificationsConfig,
-		onChainConfig:           onChainConfig,
-		userPreferences:         userPreferences,
-		riverRegistry:           riverRegistry,
-		nodes:                   nodes,
-		listener:                listener,
-		streamsTracker:          tracker,
-		metrics:                 metrics,
-		sessionTokenSigningKey:  key,
-		sessionTokenSigningAlgo: notificationsConfig.Authentication.SessionToken.Key.Algorithm,
-	}, nil
+	return service, nil
 }
 
 func (s *Service) Start(ctx context.Context) {
