+++ core/node/rpc/server.go
@@ -757,7 +757,7 @@ func (s *Service) initNotificationHandlers() error {
 	ii = append(ii, s.NewMetricsInterceptor())
 	ii = append(ii, NewTimeoutInterceptor(s.config.Network.RequestTimeout))
 
-	authInceptor, err := notifications.NewAuthenticationInterceptor(
+	authInceptor, err := auth.NewAuthenticationInterceptor(
 		s.config.Notifications.Authentication.SessionToken.Key.Algorithm,
 		s.config.Notifications.Authentication.SessionToken.Key.Key,
 	)
